Add css code to style header and under header content
Found a clever way to ensure the header logo/links stays aligned with
the boundaries of the website info container and image, by setting the
max width's of the text info and img div as a variable, I can then
calculate the addition of that variable to itself as the max width
of the header. All this to say that if I ever change the size of
the website info container the header will follow suit without me
having to do anything. Probably over complicated for this but I found
it cool!